API Documentation

K8sClusterManagers

API

native

K8sClusterManagers.addprocs_pod — Function.

addprocs_pod(np::Int;
             configure=identity,
             namespace::String="default",
             image=nothing,
             memory::String="4Gi",
             cpu::String="1",
             retry_seconds::Int=180,
             exename=`julia`,
             exeflags=``,
             params...)

Launch and connect to np worker processes. This may launch and return fewer pids than requested, if requested workers do not come online within retry_seconds.

If image is unspecified

if the julia caller is in a pod containing a single container, its image will be used for the pods as well.

if there are multiple containers in the driver's pod, an error is thrown.

For more advanced configuration, configure will be applied to the Kuber.jl pod specs of worker nodes before applying them.

source

K8sClusterManagers.default_pod — Function.

default_pod(ctx, port, cmd::Cmd, driver_name::String; image=nothing, memory::String="4Gi", cpu::String="1", base_obj=kuber_obj(ctx, empty_pod))

Kuber object representing a pod with a single worker container.

If isnothing(image) == true, the driver pod is required to have a single container, whose image will be used.

source

Missing docstring.

Missing docstring for K8sNativeManager. Check Documenter's build log for details.