API

K8sClusterManagers.K8sClusterManager — Type

K8sClusterManager(np::Integer; kwargs...)

A cluster manager using Kubernetes (k8s) which spawns additional pods as workers. Attempts to spawn np workers but may launch with less workers if the cluster has less resources available.

Arguments

np: Desired number of worker pods to be launched.

Keywords

namespace: the Kubernetes namespace to launch worker pods within. Defaults to current_namespace().

manager_pod_name: the name of the manager pod. Defaults to ENV["HOSTNAME"] which is the name of the pod when executed inside of a Kubernetes pod.

image: Docker image to use for the workers. Defaults to using the image of the Julia caller if running within a pod using a single container otherwise is a required argument.

cpu: CPU resources requested for each worker. Defaults to 1,

memory: Memory resource requested for each worker in bytes. Requests may provide a unit suffix (e.g. "G" for Gigabytes and "GiB" for Gibibytes). Defaults to "4Gi".

pending_timeout: The maximum number of seconds to wait for a "Pending" worker pod to enter the "Running" phase. Once the timeout has been reached the manager will continue with the number of workers available (<= np). Defaults to 180 (3 minutes).

configure: A function which allows modification of the worker pod specification before their creation. Defaults to identity.

K8sClusterManagers.isk8s — Function

isk8s() -> Bool

Predicate for testing if the current process is running within a Kubernetes (K8s) pod.